Handle the case of an empty lint.
There are some cases where we don't know whether there will be a need to
lint until we have started to compute the lint. Examples include commit
messages with notes to skip ci explicitly and PRs closed during linting.
To handle the case of an empty lint, we updated the code and tests to
check to make sure the lint did happen before proceeding further. This
will help prepare for the possibility of a lint being skipped while in
progress.