Force Python 3.5
As it appears some dependencies are not correctly satisfied on Python
3.6 and they are causing us some issues with getting proper
re-renderings, switch to the latest Miniconda using Python 3.5 and pin
the Python, during upgrade, to 3.5. At least on Python 3.5, we are
confident that we have all the dependencies needed to update the status
page and such. We can always revert this pinning once the underlying
issues are fixed.

[ci skip]