Add a webservice for updating feedstocks
Provides an asynchronous webservice, which listens to push events on the
feedstocks. When a push event occurs on a feedstock, the webservice
updates the commit that the appropriate submodule points to in the
feedstocks repo. This way it is a lot easier to handle updates of
various feedstocks throughout the organization on demand instead of
running a massive batch script over all of the feedstocks (even when
most haven't changed).