Use a different token for feedstock updates
Instead of using the token that is associated with the linter use a
token associated with the coordinator. This matches more closely to the
behavior that was present when feedstock updates were performed by the
webpage repo's CI. Should help mitigate rate limiting for the linter and
other services due to feedstock updates.