Drop CI skip from feedstocks commit message
Previously we had looked at having the feedstocks repo update itself
with Travis CI. However that was abandoned long ago in favor the
feedstocks webservice, which we now have. Recently disabled Travis CI on
the feedstocks repo and closed any PRs that relied on it. Given this,
there is no further need to try and disable CI in commit messages pushed
to the feedstocks repo. Hence we drop it from our commit message here.