Avoid calling GitHub API when updating feedstocks
Based on the URLs of the other feedstocks in `.gitmodules` both before
the webservice and after, they all seem to have the same format. It is
pretty easy to create a correctly formatted URL for updating these
feedstock modules without using the GitHub API. Hence we drop the
unneeded GitHub API calls and format the feedstock URL ourselves.