Name feedstock updated
Instead of simply say that we updated feedstock submodules, which is a
bit generic and not totally accurate, specify which feedstock we updated
by name. This should make it easier to see what has been updated. Also
it will be easier to catch if anything was not updated when it should
have been.