Update feedstocks listing on webpage update
If the webpage repo is updated, also update the feedstocks listing.
This is done because the feedstocks listing template could be updated in
the webpage repo, which could result in changes to the feedstocks
listing. Hence we should try regenerating the feedstocks listing in this
case so as to make any needed changes to the format.