Stage changes and commit if the index is dirty
Make sure to stage changes first and only create a commit if there are
changes staged in the index (not if there are ones in the working tree).
It is possible to have changes that can't easily be staged if there are
formatting issues (e.g. `.gitattributes` related changes) that need to
be addressed in a feedstock for instance. This provides some extra
assurance against creating empty commits.