Wait up to 1 second before linting
Appears the webhooks for PRs fire a bit in advance of when the merge ref
shows up. While this was solved by sleeping for 0.1 seconds at a time,
it appears that is no longer enough. This sleeps for 1 second at a time
in the hope that will be enough time for the merge ref to show up.