Simply checkout feedstock's remote's latest
Instead of trying to checkout the specific remote commit and move to the
`master` branch to fast forward. Just checkout the latest remote commit
to start with. This already will be restricted to `master` based on how
existing git submodules are added to the `feedstocks` repo and how we
add new ones. As a result, this will be resilient to force pushes on the
upstream remote. Also this is simpler to maintain.