Enumerate each package on a separate line
When writing out `pkg_versions.json`, write each one to a separate line.
This should keep the diffs cleaner and easier to read.

Also write the commit message so that package versions are in a bullet
point list in the body of the message. This will make it easier to
follow what was changed and keep the commit messages more readable.