Create PRs as draft
Initially create PRs as draft. Once the bot has finished prepping them,
the bot can come back and mark them ready for review.

This will make it clearer that the bot is still doing work before the PR
is ready for review/merge. Also draft PRs don't request reviews (and
depending on notification settings won't ping people initially). After
the bot is done prepping the PR, it can come back and mark it as ready
for review, which will request reviews from maintainers and generate the
usual level of notifications.