Visually distinguish between hints and lints
As I explain [here](https://github.com/conda-forge/conda-smithy/pull/2126#discussion_r1834476654) it can be really easy to miss the distinction between hints and linting errors. (The difference is indicated in a big block of boilerplate text.) Using :information_source: for hints and :x: for linter errors, I hope to make this easier to understand the severity at a glance.